feat: adding opt in
