+++ src/testing/asset.ts
@@ -1,6 +1,7 @@
 import { Account, Algodv2, Kmd, makeAssetCreateTxnWithSuggestedParamsFromObject } from 'algosdk'
 import { algos, microAlgos } from '../amount'
-import { ensureFunded, transferAsset } from '../transfer'
+import { optIn } from '../asset'
+import { ensureFunded } from '../transfer'
 
 export async function generateTestAsset(client: Algodv2, sender: Account, total?: number) {
   total = !total ? Math.floor(Math.random() * 100) + 20 : total
@@ -34,19 +35,6 @@ export async function generateTestAsset(client: Algodv2, sender: Account, total?
   const assetId = ptx['asset-index']
 
   return assetId
-}
-
-export async function optIn(algod: Algodv2, account: Account, assetId: number) {
-  await transferAsset(
-    {
-      from: account,
-      to: account.addr,
-      assetId,
-      amount: 0,
-      note: `Opt in asset id ${assetId}`,
-    },
-    algod,
-  )
 }
 
 export async function ensureFundsAndOptIn(algod: Algodv2, account: Account, assetId: number, kmd: Kmd) {
